Moved back into service. Was moved temporarily for testing purposes in an earlier commit
