+++ backend/src/main/java/no/ntnu/database/controllers/AuthenticationController.java
@@ -4,6 +4,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import java.security.Principal;
+
+import no.ntnu.database.services.TwoFactorRequests;
 import no.ntnu.dto.TwoFactorDto;
 import no.ntnu.security.SecurityUtil;
